Parent home components pull data
