+++ src/electionguard_gui/web/components/guardian/guardian-home-component.js
@@ -1,14 +1,18 @@
 import KeyCeremonyList from "../shared/key-ceremony-list-component.js";
 import DecryptionList from "./decryption-list-component.js";
+import Spinner from "../shared/spinner-component.js";
 
 export default {
   components: {
     KeyCeremonyList,
     DecryptionList,
+    Spinner,
   },
   data() {
     return {
+      loading: true,
       decryptions: [],
+      keyCeremonies: [],
     };
   },
   methods: {
@@ -23,21 +27,39 @@ export default {
         console.error(result.error);
       }
     },
+    refreshKeyCeremonies: async function () {
+      const result = await eel.get_key_ceremonies()();
+      if (result.success) {
+        this.keyCeremonies = result.result;
+      } else {
+        console.error(result.error);
+      }
+    },
   },
   async mounted() {
     eel.expose(this.keyCeremoniesChanged, "key_ceremonies_changed");
     console.log("begin watching for key ceremonies");
     eel.watch_key_ceremonies();
-    await this.$refs.keyCeremonyListComponent.refreshKeyCeremonies();
+    await this.refreshKeyCeremonies();
     await this.refreshDecryptions();
+    this.loading = false;
   },
   unmounted() {
     console.log("stop watching key ceremonies");
     eel.stop_watching_key_ceremonies();
   },
   template: /*html*/ `
-  <h1>Guardian Home</h1>
-  <key-ceremony-list :show-when-empty="true" :is-admin="false" ref="keyCeremonyListComponent"></key-ceremony-list>
-  <decryption-list :decryptions="decryptions"></decryption-list>
+  <div class="container">
+    <h1>Guardian Home</h1>
+    <spinner :visible="loading"></spinner>
+    <div v-if="!loading" class="row">
+      <div class="col-12 col-lg-6">
+        <key-ceremony-list :show-when-empty="true" :is-admin="false" :key-ceremonies="keyCeremonies"></key-ceremony-list>
+      </div>
+      <div class="col-12 col-lg-6">
+        <decryption-list :decryptions="decryptions"></decryption-list>
+      </div>
+    </div>
+  </div>
   `,
 };
